bug fix - build command was writing dists files to the wrong directory when there is originally-shared-dir
